Added placeholder for Template-Based (CableLabs) and CMP (RFC 4210)
Added placeholder for two Provisioning/Re-Provisioning protocols that might be worth taking in consideration. The first one is the Template-Based Credentials Re-Provisioning Protocol that has been developed internally and that we would like to standadize (message-based). The second one is the Certificate Management Protocol which, in its easiest implementation, might provide a method that uses messages already implemented in crypto libraries.